Added placeable bomb

+++ Shared/Bomb.cs
@@ -10,9 +10,9 @@ namespace BomberGopnik.Shared
     {
         public int Timer { get; set; }
         public int Radius { get; set; }
-        public int StartX { get; set; }
+        public double StartX { get; set; }
         public int Length { get; set; }
-        public int StartY { get; set; }
+        public double StartY { get; set; }
         public Bomb(int timer, int radius, int startX, int length, int startY)
         {
             Timer = timer;
